feat: 내정보 컨텐츠 섹션 에러바운더리 적용, 내정보 - 내가 썼어요 컨텐츠 컴포넌트 분리
